fix: update style and tooltips of asset widgets (#1309)
* chore: set min-height to Severity... widget


* feat: add `setXYSharedTooltipTextWithRate` to xy-chart-helper


* chore: use `setXYSharedTooltipTextWithRate` to widgets


* chore: manage minor style issue


---------
