+++ apps/web/src/common/composables/amcharts5/xy-chart-helper.ts
@@ -5,7 +5,7 @@ import type { IDateAxisSettings, IXYChartSettings, IXYSeriesSettings } from '@am
 import * as am5xy from '@amcharts/amcharts5/xy';
 import bytes from 'bytes';
 
-import { byteFormatter } from '@cloudforet/core-lib';
+import { byteFormatter, commaFormatter } from '@cloudforet/core-lib';
 
 import type { Currency } from '@/store/modules/display/config';
 import type { CurrencyRates } from '@/store/modules/display/type';
@@ -251,6 +251,28 @@ export const setXYSharedTooltipTextByUsage = (chart: am5xy.XYChart, tooltip: am5
             value = bytes.parse(`${value}${sourceUnit}`);
             value = byteFormatter(value);
             _text += `\n[${s.get('stroke')?.toString()}; fontSize: 10px]●[/] [fontSize: 14px;}]${s.get('name')}:[/] [bold; fontSize: 14px]${value}[/]`;
+        });
+        return _text;
+    });
+};
+export const setXYSharedTooltipTextWithRate = (chart: am5xy.XYChart, tooltip: am5.Tooltip): void => {
+    tooltip.label.adapters.add('text', (text, target) => {
+        let totalValue = 0;
+        const seriesList: any[] = []; // { color: string, name: string, value: number }[]
+        chart.series.each((s) => {
+            const fieldName = s.get('valueYField') || s.get('valueXField') || '';
+            const value = target.dataItem?.dataContext?.[fieldName];
+            totalValue += value;
+            seriesList.push({
+                color: s.get('stroke')?.toString() ?? '',
+                name: s.get('name') ?? '',
+                value: value ?? 0,
+            });
+        });
+        let _text = `Total: [bold; fontSize: 14px]${commaFormatter(totalValue)}[/]`;
+        seriesList.forEach((s) => {
+            const rate = Math.round((s.value / totalValue) * 100);
+            _text += `\n[${s.color}; fontSize: 10px]●[/] [fontSize: 14px;}]${s.name}:[/] [bold; fontSize: 14px]${commaFormatter(s.value)}[/] (${rate}%)`;
         });
         return _text;
     });
