+++ apps/web/src/services/dashboards/widgets/asset-widgets/severity-status-by-service/SeverityStatusByServiceWidget.vue
@@ -4,18 +4,26 @@
                   @refresh="refreshWidget"
     >
         <div class="data-container">
-            <div class="chart-wrapper"
-                 :style="{'grid-template-columns': `repeat(auto-fill, ${state.boxWidth-4}px)`}"
+            <p-data-loader class="chart-wrapper"
+                           :loading="state.loading"
+                           :data="state.refinedData"
+                           loader-type="skeleton"
+                           :loader-backdrop-opacity="1"
+                           show-data-from-scratch
             >
-                <div v-for="(data, idx) in state.refinedData"
-                     :key="`box-${idx}`"
-                     v-tooltip.bottom="`${data.service}: ${data.value}`"
-                     class="status-box"
-                     :style="{'background-color': SEVERITY_STATUS_MAP[data.severity].color}"
+                <div class="box-wrapper"
+                     :style="{'grid-template-columns': `repeat(auto-fill, ${state.boxWidth-4}px)`}"
                 >
-                    <span class="text">{{ data.service }}</span>
+                    <div v-for="(data, idx) in state.refinedData"
+                         :key="`box-${idx}`"
+                         v-tooltip.bottom="`${data.service}: ${data.value}`"
+                         class="status-box"
+                         :style="{'background-color': SEVERITY_STATUS_MAP[data.severity].color}"
+                    >
+                        <span class="text">{{ data.service }}</span>
+                    </div>
                 </div>
-            </div>
+            </p-data-loader>
             <div class="legend-wrapper">
                 <div v-for="status in SEVERITY_STATUS_MAP_VALUES"
                      :key="`status-${status.label}`"
@@ -38,6 +46,7 @@ import {
     computed, defineExpose, defineProps, reactive, toRefs,
 } from 'vue';
 
+import { PDataLoader } from '@spaceone/design-system';
 import dayjs from 'dayjs';
 import { flattenDeep, min } from 'lodash';
 
@@ -209,13 +218,17 @@ defineExpose<WidgetExpose>({
 .severity-status-by-service {
     .data-container {
         .chart-wrapper {
-            display: grid;
-            grid-auto-flow: row;
-            gap: 0.25rem;
-            .status-box {
-                height: 4.5rem;
-                font-weight: 500;
-                padding: 0.5rem;
+            min-height: 16.5rem;
+            .box-wrapper {
+                height: 100%;
+                display: grid;
+                grid-auto-flow: row;
+                gap: 0.25rem;
+                .status-box {
+                    height: 4.5rem;
+                    font-weight: 500;
+                    padding: 0.5rem;
+                }
             }
         }
         .legend-wrapper {
